🚚 rename: document 경로 수정
entity 패키지가 생기며, domain 패키지 안으로 들어감
